core: Refactor a lot of core/

+++ base/rand/rand.c
@@ -1,13 +1,11 @@
-#include <stdlib.h>
 #include "rand.h"
-// rain_rand()
-// rain_srand()
+#include <stdlib.h>
 
 void rain_rand(box * ret) {
-    double num_rand = (double)rand()  / (double)RAND_MAX;
-    rain_set_float(ret, (double)num_rand);
+  double num_rand = (double)rand()  / (double)RAND_MAX;
+  rain_set_float(ret, num_rand);
 }
 
 void rain_srand(box * ret, box * seed) {
-    srand(seed->data.ui);
+  srand(seed->data.ui);
 }
